Fix fatal runtime panic passing invalid pointer
The last argument to hs_scan is "void*", a pointer to an abitrary
context. "hsScan" incorrectly uses "unsafe.Pointer(h)" when h is not a
pointer. This can cause fatal runtime panic when Go passes arguments
across the Go <-> C boundaries. This does not happen all the time, as it
usually "looks" like a pointer.

In reality, h is an unsigned long value belonging to "handles". We can
coerce this into "void*" on the C side with the added "safe_hs_scan"